Redux-ify the widget more
