+++ source/_patterns/02-molecules/example-widget/demo/example-widgets.twig
@@ -6,3 +6,7 @@
 
   {% include '@molecules/example-widget/_example-widget.twig' %}
 {% endblock column_1 %}
+
+{% block column_2 %}{% endblock column_2 %}
+
+{% block column_3 %}{% endblock column_3 %}
