+++ source/_patterns/02-molecules/example-widget/src/store.js
@@ -1,0 +1,8 @@
+import createStore from '../lib/redux-lite';
+
+import reducer from './reducer';
+
+// Wire up our reducer to our state
+const store = createStore(reducer);
+
+export default store;
